Inquire Widget format
Adjusted the spacing the of the widget to allow for a border 9-slice. This still gives some erratic behaviour but works in most scenarios. Will need to explore if this works on different resolution screens as the border image layout depends on the scale of the widget in relation to the player camera. The effect is that getting closer to the widget zooms away from the texture used in the 9-slice.